llm-sort-floating: create new(!) experiment setup for double arrays
ChatGPT and Gemini told me that Radix sort can't be implemented
for floating points. I didn't pressure them, because I think they
will be able to come up with an implementation if I insist. This
makes me question about the usefulness of this experiment as we
are not really interrogating them.

Claude says her algorithm works for only positive numbers. By
default I used positive numbers, negatives can be tested as well.

Ideally, this LLM's algorithm can sort this, can't sort this type
of experiment evaluation is needed. As we are not entirely testing
these LLMs' creativity, I didn't want to spend time on setting up
a new environment for this. Yet, this could be easily achieved
manually. Just make sure `SORT_VERIFICATION` flag is set to True.
This flag is true by default in this experiment.