all: introduce debug-measure envs, compile time experiment flags
@ikolomiko is a clown because he doesn't put new lines at the end
of his files. Don't be like him.

Hardcoded arrays can be defined like this too maybe. This could be
useful because R3 and R4 Minima's memories are not alike. For some
experiments we can go with bigger arrays, but this time we need to
adapt its content (because -300:+300 ain't cool when you have 1024
elements in your array). I don't know maybe -300:+300 may not be
hard requirement. I wasn't really good at data types in computers
at that time.